refactor: put all code to save a file into FirebaseUploadService class
